fix : DB 테이블 구조 전체 점검(id LONG형, 필요 없는 assinged컬럼 삭제)

+++ weddingyou_backend/src/main/java/com/mysite/weddingyou_backend/plannerLogin/PlannerLogin.java
@@ -24,7 +24,7 @@ public class PlannerLogin {
 	@Id // pk 지정
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "planner_id")
-	private int plannerId;
+	private Long plannerId;
 
 	@Column(name = "name", nullable = false)
 	private String name;
